Update to bottom navigation and add saved area list

+++ app/src/main/java/com/chrisa/covid19/features/area/presentation/AreaFragment.kt
@@ -118,13 +118,13 @@ class AreaFragment : Fragment(R.layout.fragment_area) {
                 )
             )
 
-            currentInfectionRate.text = NumberFormatter.format(areaCasesModel.currentInfectionRate)
+            currentInfectionRate.text = NumberFormatter.format(areaCasesModel.currentInfectionRate.toInt())
             infectionRateChangeThisWeek.text =
-                NumberFormatter.getChangeText(areaCasesModel.changeInInfectionRatesThisWeek)
+                NumberFormatter.getChangeText(areaCasesModel.changeInInfectionRatesThisWeek.toInt())
             infectionRateChangeThisWeek.setTextColor(
                 ContextCompat.getColor(
                     changeInNewCasesThisWeek.context,
-                    NumberFormatter.getChangeColour(areaCasesModel.changeInInfectionRatesThisWeek)
+                    NumberFormatter.getChangeColour(areaCasesModel.changeInInfectionRatesThisWeek.toInt())
                 )
             )
